Added post policy && set comment using CTE recursive

+++ app/Http/Controllers/API/PostController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Post\PostCreateRequest;
 use App\Http\Resources\Post\PostIndexResource;
 use App\Http\Resources\Post\PostResource;
+use App\Models\Comment;
 use App\Models\Post;
 use Illuminate\Http\Request;
 
@@ -31,13 +32,19 @@ class PostController extends Controller
 
     public function update(PostCreateRequest $request, Post $post)
     {
+        $this->authorize('update', $post);
+
         return new PostResource(tap($post)->update($request->validated()));
     }
 
     public function destroy(Post $post)
     {
+        $this->authorize('update', $post);
+
         $post->comments()->delete();
         $post->likes()->delete();
         $post->delete();
+
+        return response()->json(['message' => 'Post deleted']);
     }
 }
